nice

+++ code/modules/power/singularity/particle_accelerator/particle.dm
@@ -58,11 +58,12 @@
 	M.rad_act(energy*6)
 
 /obj/effect/accelerated_particle/proc/move()
+	if(QDELETED(src))
+		return
 	if(!step(src,dir))
 		forceMove(get_step(src,dir))
 	movement_range--
 	if(movement_range == 0)
 		qdel(src)
 	else
-		sleep(speed)
-		move()
+		addtimer(CALLBACK(src, .proc/move), speed)
